impr: split index.d.ts into multiple definition file
