+++ typings/bitcore.d.ts
@@ -1,0 +1,13 @@
+/**
+ * @class bitcore
+ */
+export class bitcore {
+}
+
+export namespace bitcore {
+    /**
+     * @constructor
+     */
+    class Error {
+    }
+}
