+++ typings/buffer/BufferReader.d.ts
@@ -1,0 +1,14 @@
+/**
+ *
+ * @param buf
+ * @returns {BufferReader}
+ * @constructor
+ */
+export class BufferReader {
+    constructor(buf: any);
+
+    /**
+     * reads a length prepended buffer
+     */
+    readVarLengthBuffer(): void;
+}
