+++ typings/buffer/BufferWriter.d.ts
@@ -1,0 +1,9 @@
+/**
+ *
+ * @param obj
+ * @returns {BufferWriter}
+ * @constructor
+ */
+export class BufferWriter {
+    constructor(obj: any);
+}
