+++ typings/crypto/BN.d.ts
@@ -1,0 +1,5 @@
+/**
+ * @class BN
+ */
+export class BN {
+}
