+++ typings/govobject/Proposal.d.ts
@@ -1,0 +1,8 @@
+/**
+ * Represents 'proposal' Governance Object
+ *
+ * @constructor
+ */
+export class Proposal {
+    constructor();
+}
