+++ typings/govobject/Trigger.d.ts
@@ -1,0 +1,13 @@
+
+
+
+/**
+ * Represents 'trigger' Governance Object
+ *
+ * @constructor
+ */
+export class Trigger {
+    constructor();
+}
+
+
