+++ typings/transaction/Output.d.ts
@@ -1,0 +1,8 @@
+/**
+ * @param args
+ * @returns {Output}
+ * @constructor
+ */
+export class Output {
+    constructor(args: any);
+}
