+++ typings/transaction/input/MultiSigScriptHashInput.d.ts
@@ -1,0 +1,8 @@
+
+
+/**
+ * @constructor
+ */
+export class MultiSigScriptHashInput {
+    constructor();
+}
